avoid php error when file do not exist fix #95
just avoid the error. there is a temptation to try to use a different
file but maybe if someone will complain in the future. right now it
seems like to much work to support people that removed files they should
not have removed, or not done it in the proper way